后端框架已搭好

+++ src/main/java/teektok/dto/analysis/VideoAnalysisVO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Schema(description = "视频行为统计数据")
 public class VideoAnalysisVO {
-    private Integer playCount;
-    private Integer likeCount;
-    private Integer commentCount;
+    private Long playCount;
+    private Long likeCount;
+    private Long commentCount;
 }
